feat(message): support useCredentials and custom headers
